add classchecker

+++ 275-jvm-test/src/main/java/com/njuse/jvmfinal/instructions/conversion/I2B.java
@@ -12,6 +12,6 @@ public class I2B extends NoOperandsInstruction {
     @Override
     public void execute(StackFrame frame) {
         int b=frame.getOperandStack().popInt();
-        frame.getOperandStack().pushInt(b&0xFF);
+        frame.getOperandStack().pushInt((byte)b&0xFF);
     }
 }
